MachO: add wrapper struct
Add a wrapper struct MachO which encapsulates all the relevant Mach-O
related fields in one place [for now, that is just the header].

The raw Mach-O structs and data are placed throughout the binary file,
so it is cumbersome to deal with outside of the parsing code. Therefore,
after parsing, we encapsulate the parsed structs into a more convenient
hierarchy of structs.

* Add MachO struct in pkg/obj/machO/parser.go
* Refactor and move logic from pkg/obj/machO/machO.go
* Update usages in internal/loader.go